+++ source/utils/BUILD
@@ -6,6 +6,7 @@ cc_library(
     "base64.h",
     "random_generator.h",
     "circular_buffer.h",
+    "alarm.h",
   ],
   srcs = [
     "random_generator.cc",
